+++ android/src/test/kotlin/com/mrrhak/screen_lock_detector/ScreenLockDetectorPluginTest.kt
@@ -5,23 +5,26 @@ import io.flutter.plugin.common.MethodChannel
 import org.mockito.Mockito
 import kotlin.test.Test
 
-/*
- * This demonstrates a simple unit test of the Kotlin portion of this plugin's implementation.
- *
- * Once you have built the plugin's example app, you can run these tests from the command
- * line by running `./gradlew testDebugUnitTest` in the `example/android/` directory, or
- * you can run them directly from IDEs that support JUnit such as Android Studio.
- */
-
 internal class ScreenLockDetectorPluginTest {
     @Test
-    fun onMethodCall_getPlatformVersion_returnsExpectedValue() {
+    fun onMethodCall_checkScreenStatus_returnsUnknownWhenNotAttached() {
         val plugin = ScreenLockDetectorPlugin()
 
-        val call = MethodCall("getPlatformVersion", null)
+        val call = MethodCall("checkScreenStatus", null)
         val mockResult: MethodChannel.Result = Mockito.mock(MethodChannel.Result::class.java)
         plugin.onMethodCall(call, mockResult)
 
-        Mockito.verify(mockResult).success("Android " + android.os.Build.VERSION.RELEASE)
+        Mockito.verify(mockResult).success("UNKNOWN")
+    }
+
+    @Test
+    fun onMethodCall_unknownMethod_returnsNotImplemented() {
+        val plugin = ScreenLockDetectorPlugin()
+
+        val call = MethodCall("unknownMethod", null)
+        val mockResult: MethodChannel.Result = Mockito.mock(MethodChannel.Result::class.java)
+        plugin.onMethodCall(call, mockResult)
+
+        Mockito.verify(mockResult).notImplemented()
     }
 }