+++ lib/src/screen_lock_detector_platform_interface.dart
@@ -3,7 +3,7 @@ import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'screen_lock_detector_method_channel.dart';
 import 'common/screen_status.dart';
 
-/// The interface that implementations of screen_lock_detector
+/// The interface that implementations of screen_lock_detector must implement.
 abstract class ScreenLockDetectorPlatform extends PlatformInterface {
   /// Constructs a ScreenLockDetectorPlatform.
   ScreenLockDetectorPlatform() : super(token: _token);