+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.1
+- Add FlutterFramework dependency and improve tests in iOS platform
+- Updates minimum supported SDK version to Flutter 3.44/Dart 3.12
+- Migrates Android to built-in Kotlin
+
 ## 1.0.0
 - Detect when the screen is locked
 - Detect when the screen is unlocked